refactor: api 호출 query파일에 병합
