--- src/apis/mypage/deletePostComment.ts
@@ -1,8 +1,0 @@
-import axiosInstance from "@/apis/utils/axiosInstance";
-
-const deletePostComment = async (commentId: number) => {
-  const res = await axiosInstance.delete(`/mypage/comments/${commentId}`);
-  return res.data.result;
-};
-
-export default deletePostComment;
