--- src/apis/mypage/fetchPostReview.ts
@@ -1,8 +1,0 @@
-import axiosInstance from "@/apis/utils/axiosInstance";
-
-const fetchPostReview = async () => {
-  const res = await axiosInstance.get("/mypage/reviews");
-  return res.data.data || [];
-};
-
-export default fetchPostReview;
