+++ src/queries/mypage/useDeletePostCommentMutation.ts
@@ -1,5 +1,10 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import deletePostComment from "@/apis/mypage/deletePostComment";
+import axiosInstance from "@/apis/utils/axiosInstance";
+
+const deletePostComment = async (commentId: number) => {
+  const res = await axiosInstance.delete(`/mypage/comments/${commentId}`);
+  return res.data.result;
+};
 
 export const useDeletePostCommentMutation = () => {
   const queryClient = useQueryClient();
