+++ src/queries/mypage/useLogoutMutation.ts
@@ -1,6 +1,10 @@
 import { useMutation } from "@tanstack/react-query";
-import fetchLogout from "@/apis/mypage/fetchLogout";
+import axiosInstance from "@/apis/utils/axiosInstance";
 
+const fetchLogout = async (): Promise<boolean> => {
+  const response = await axiosInstance.post("/users/auth/logout", undefined);
+  return response.status === 204;
+};
 export const useLogoutMutation = (options?: {
   onSuccess?: (data: boolean) => void;
   onError?: (error: unknown) => void;
