+++ src/queries/mypage/usePostCommentQuery.ts
@@ -1,5 +1,10 @@
 import { useQuery } from "@tanstack/react-query";
-import fetchPostComment from "@/apis/mypage/fetchPostComment";
+import axiosInstance from "@/apis/utils/axiosInstance";
+
+const fetchPostComment = async () => {
+  const res = await axiosInstance.get("/mypage/comments");
+  return res.data.data || [];
+};
 
 export const usePostCommentQuery = () =>
   useQuery({
