+++ src/queries/mypage/usePostReviewQuery.ts
@@ -1,5 +1,10 @@
 import { useQuery } from "@tanstack/react-query";
-import fetchPostReview from "@/apis/mypage/fetchPostReview";
+import axiosInstance from "@/apis/utils/axiosInstance";
+
+const fetchPostReview = async () => {
+  const res = await axiosInstance.get("/mypage/reviews");
+  return res.data.data || [];
+};
 
 export const usePostReviewQuery = () =>
   useQuery({
