+++ src/queries/mypage/useUpdateProfileImageMutation.ts
@@ -1,5 +1,16 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import fetchUpdateProfileImage from "@/apis/mypage/fetchUpdateProfileImage";
+import axiosInstance from "@/apis/utils/axiosInstance";
+
+const fetchUpdateProfileImage = async (file: File): Promise<string> => {
+  const formData = new FormData();
+  formData.append("profileImageUrl", file);
+  const response = await axiosInstance.post("/mypage/profile-image", formData, {
+    headers: {
+      "Content-Type": "multipart/form-data",
+    },
+  });
+  return response.data.data.profileImageUrl;
+};
 
 export const useUpdateProfileImageMutation = (options?: {
   onSuccess?: (data: string) => void;
