+++ src/queries/mypage/useWithdrawMutation.ts
@@ -1,5 +1,12 @@
 import { useMutation } from "@tanstack/react-query";
-import fetchWithdraw from "@/apis/mypage/fetchWithdraw";
+import axiosInstance from "@/apis/utils/axiosInstance";
+
+const fetchWithdraw = async (reason: string): Promise<boolean> => {
+  const response = await axiosInstance.delete("/mypage/withdraw", {
+    data: { withdrawReason: reason },
+  });
+  return response.status === 200;
+};
 
 export const useWithdrawMutation = (options?: {
   onSuccess?: (data: boolean) => void;
